Reorganized project structure. Optimized code and fix login form display for desktop environments.
