+++ lib/app_controller.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
-import 'package:k_front/pages/desktop/signup_form.dart';
-import 'package:k_front/pages/tablet/login_form.dart';
+import 'package:k_front/routes/pages/common/login_page.dart';
 import 'package:k_front/routes/index.dart';
+import 'package:k_front/routes/pages/common/signup_page.dart';
 import 'package:k_front/themes/theme_data.dart';
 import 'package:k_front/themes/theme_provider.dart';
 import 'package:k_front/utils/device_checker.dart';
@@ -50,8 +50,8 @@ class KCalFront extends StatelessWidget {
           '/': (context) => const WebIndex(),
           // 'desktop/login': (context) => const DesktopLoginForm(),
           // 'mobile/login': (context) => const MobileLoginForm(),
-          'tablet/login': (context) => const TabletLoginForm(),
-          'signup': (context) => const DesktopSignUpPage(),
+          'login': (context) => const LoginPage(),
+          'signup': (context) => const SignUpPage(),
         },
       ),
     );
