+++ lib/model/logged_user.dart
