+++ lib/routes/index.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:k_front/pages/desktop/index_page.dart';
-import 'package:k_front/pages/tablet/index_page.dart';
+import 'package:k_front/routes/pages/desktop/index_page.dart';
+import 'package:k_front/routes/pages/tablet/index_page.dart';
 
 import '../controller/impl/api_user_controller_impl.dart';
 import '../controller/api_user_controller.dart';
-import '../pages/mobile/index_page.dart';
+import 'pages/mobile/index_page.dart';
 import '../utils/device_checker.dart';
 
 class WebIndex extends StatefulWidget {
