+++ lib/routes/pages/tablet/index_page.dart
@@ -1,17 +1,14 @@
 import 'package:flutter/material.dart';
-import 'package:k_front/pages/desktop/login_form.dart';
-import 'package:k_front/pages/tablet/signup_form.dart';
 
-import '../../themes/theme_data.dart';
-import '../../utils/device_checker.dart';
-import '../../widgets/common/app_bar.dart';
-import '../../widgets/desktop/background_index.dart';
-import '../../widgets/desktop/carrousel_index.dart';
-import 'signup_form.dart';
+import '../../../themes/theme_data.dart';
+import '../../../widgets/common/app_bar.dart';
+import '../../../widgets/common/background_index.dart';
+import '../../../widgets/desktop/carrousel_index.dart';
+import 'package:k_front/routes/pages/common/signup_page.dart';
 
 const title = 'kCal Control';
 
-Scaffold desktopIndex(BuildContext context) {
+Scaffold tabletIndex(BuildContext context) {
   return Scaffold(
     appBar: DAppBar(
       title: title,
@@ -23,9 +20,8 @@ Scaffold desktopIndex(BuildContext context) {
                 onPressed: () {
                   Navigator.push(
                       context,
-                      //TODO: adapt desktop signup page or unify it with tablet signup page
                       MaterialPageRoute(
-                          builder: (context) => const TabletSignUpPage()));
+                          builder: (context) => const SignUpPage()));
                 },
                 child: Text('Sign up',
                     style: TextStyle(
@@ -33,6 +29,19 @@ Scaffold desktopIndex(BuildContext context) {
                             .textTheme
                             .headlineLarge
                             ?.color)))),
+        Container(
+            margin: const EdgeInsets.only(right: 20),
+            child: TextButton(
+                onPressed: () {
+                  Navigator.pushNamed(context, 'login');
+                },
+                style: ButtonStyle(
+                    backgroundColor: WidgetStateProperty.all<Color>(
+                        Theme.of(context).highlightColor)),
+                child: Text('Log in',
+                    style: TextStyle(
+                        color:
+                            Theme.of(context).textTheme.headlineLarge?.color))))
       ],
     ),
     body: Stack(
@@ -46,7 +55,7 @@ Scaffold desktopIndex(BuildContext context) {
                 children: <Widget>[
                   Center(
                     child: SizedBox(
-                        width: MediaQuery.of(context).size.width * 0.65,
+                        width: MediaQuery.of(context).size.width * 0.96,
                         height: MediaQuery.of(context).size.height * 0.20,
                         child: Center(
                           child: Text('Welcome to kCal Control',
@@ -55,26 +64,13 @@ Scaffold desktopIndex(BuildContext context) {
                   ),
                   Center(
                     child: SizedBox(
-                        width: MediaQuery.of(context).size.width * 0.6,
+                        width: MediaQuery.of(context).size.width * 0.96,
                         height: MediaQuery.of(context).size.height * 0.50,
                         child: IndexDesktopCarousel(context)),
                   )
                 ],
               ),
             ),
-            Padding(
-              padding: EdgeInsets.all(MediaQuery.of(context).size.width * 0.02),
-              child: ClipRRect(
-                borderRadius: BorderRadius.circular(10),
-                child: Container(
-                    decoration: BoxDecoration(
-                      color: Theme.of(context).primaryColor,
-                    ),
-                    width: MediaQuery.of(context).size.width * 0.25,
-                    height: 475,
-                    child: const DesktopLoginForm()),
-              ),
-            )
           ],
         ),
       ],
